Minor fixes
I Created a release for both goChem and SCU
and our go.mod no points to those.

I also fixed a minor issue in packagetools/INSTALL
which affects the latest distribution and gave a heads up
of the error in the README.md. I don't think the problem
is serious enough to create a new RC right now.